+++ Source/PluginEditor.cpp
@@ -17,21 +17,31 @@ ViiveAudioProcessorEditor::ViiveAudioProcessorEditor(ViiveAudioProcessor& p)
 	m_delayGroup.addAndMakeVisible(m_offsetKnob);
 	m_delayGroup.addAndMakeVisible(m_tempoSyncLButton);
 	m_delayGroup.addAndMakeVisible(m_tempoSyncRButton);
-	addAndMakeVisible(m_delayModeLabel);
-	m_delayModeLabel.setText("Delay Mode:", juce::NotificationType::dontSendNotification);
-	m_delayModeLabel.setEditable(false);
-	m_delayModeLabel.setSize(80, 25);
 	addAndMakeVisible(m_modeSelector);
-	m_modeSelector.addItemList(Parameters::delayModes, 1);
-	m_modeSelector.setEditableText(false);
-	m_modeSelector.setSize(130, 25);
-	m_modeSelector.onChange = [this] { modeSelectorChanged();};
-	auto* param = dynamic_cast<juce::AudioParameterChoice*>(m_audioProcessor.apvts.getParameter(delayModeParamID.getParamID()));
-	m_modeSelector.setSelectedId(param->getIndex() + 1, juce::dontSendNotification);
 	addAndMakeVisible(m_delayGroup);
 
-    m_filterGroup.setText("Filter");
+    m_filterModGroup.setText("Modulation");
+	m_filterModGroup.setTextLabelPosition(juce::Justification::horizontallyCentred);
+	m_filterModGroup.setColour(juce::GroupComponent::outlineColourId, juce::Colours::transparentBlack);
+	m_filterModGroup.addAndMakeVisible(m_lowCutModRateKnob);
+	m_filterModGroup.addChildComponent(m_lowCutModNoteKnob);
+	m_filterModGroup.addAndMakeVisible(m_lowCutModDepthKnob);
+	m_filterModGroup.addAndMakeVisible(m_lowCutModPhaseKnob);
+	m_filterModGroup.addAndMakeVisible(m_highCutModRateKnob);
+	m_filterModGroup.addChildComponent(m_highCutModNoteKnob);
+	m_filterModGroup.addAndMakeVisible(m_highCutModDepthKnob);
+	m_filterModGroup.addAndMakeVisible(m_highCutModPhaseKnob);
+	m_filterModGroup.addAndMakeVisible(m_lowCutModTempoSyncButton);
+	m_filterModGroup.addAndMakeVisible(m_highCutModTempoSyncButton);
+	m_filterModGroup.addAndMakeVisible(m_lowCutModShapeSelector);
+	m_filterModGroup.addAndMakeVisible(m_highCutModShapeSelector);
+	m_filterModGroup.addAndMakeVisible(m_lowCutLfoIndicator);
+	m_filterModGroup.addAndMakeVisible(m_highCutLfoIndicator);
+	m_filterModGroup.addAndMakeVisible(m_highCutModInvertButton);
+
+	m_filterGroup.setText("Filter");
     m_filterGroup.setTextLabelPosition(juce::Justification::horizontallyCentred);
+	m_filterGroup.addAndMakeVisible(m_filterModGroup);
 	m_filterGroup.addAndMakeVisible(m_lowCutFreqKnob);
 	m_filterGroup.addAndMakeVisible(m_lowCutQKnob);
 	m_filterGroup.addAndMakeVisible(m_highCutFreqKnob);
@@ -54,20 +64,35 @@ ViiveAudioProcessorEditor::ViiveAudioProcessorEditor(ViiveAudioProcessor& p)
 
 	addAndMakeVisible(m_meter);
 
+#if JUCE_DEBUG
+	m_lfoScope = std::make_unique<LFOScope>(p.getLFOScopeBuffer());
+	addAndMakeVisible(*m_lfoScope);
+#endif
+
 	bool syncL = m_audioProcessor.apvts.getParameter(tempoSyncLParamID.getParamID())->getValue() > 0.5f;
 	bool syncR = m_audioProcessor.apvts.getParameter(tempoSyncRParamID.getParamID())->getValue() > 0.5f;
+	bool lowCutSync = m_audioProcessor.apvts.getParameter(lowCutModTempoSyncParamID.getParamID())->getValue() > 0.5f;
+	bool highCutSync = m_audioProcessor.apvts.getParameter(highCutModTempoSyncParamID.getParamID())->getValue() > 0.5f;
 
-	updateDelayKnobs(syncL, syncR);
+	updateSyncedKnobs(syncL, syncR, lowCutSync, highCutSync);
 
 	m_audioProcessor.apvts.addParameterListener(tempoSyncLParamID.getParamID(), this);
 	m_audioProcessor.apvts.addParameterListener(tempoSyncRParamID.getParamID(), this);
-	setSize (770, 370);
+	m_audioProcessor.apvts.addParameterListener(lowCutModTempoSyncParamID.getParamID(), this);
+	m_audioProcessor.apvts.addParameterListener(highCutModTempoSyncParamID.getParamID(), this);
+#if JUCE_DEBUG
+	setSize(770, 670);
+#else
+	setSize(770, 490);
+#endif
 }
 
 ViiveAudioProcessorEditor::~ViiveAudioProcessorEditor()
 {
 	m_audioProcessor.apvts.removeParameterListener(tempoSyncLParamID.getParamID(), this);
 	m_audioProcessor.apvts.removeParameterListener(tempoSyncRParamID.getParamID(), this);
+	m_audioProcessor.apvts.removeParameterListener(lowCutModTempoSyncParamID.getParamID(), this);
+	m_audioProcessor.apvts.removeParameterListener(highCutModTempoSyncParamID.getParamID(), this);
 }
 
 //==============================================================================
@@ -77,20 +102,26 @@ void ViiveAudioProcessorEditor::paint (juce::Graphics& g)
     g.fillAll(juce::Colours::darkgrey);
 	g.setColour(juce::Colours::white);
 	g.setFont(12.0f);
-	g.drawText(JucePlugin_VersionString, getLocalBounds().reduced(5), juce::Justification::topRight);
+	g.drawText("viive 0.2.beta-1", getLocalBounds().reduced(5), juce::Justification::topRight);
+
+//	g.drawText("viive " JucePlugin_VersionString, getLocalBounds().reduced(5), juce::Justification::topRight);
 }
 
 void ViiveAudioProcessorEditor::resized()
 {
 	auto bounds = getLocalBounds();
     int y = 10;
+#ifdef JUCE_DEBUG
+	int height = bounds.getHeight() - 30 - 180;
+#else
 	int height = bounds.getHeight() - 20;
-
-    m_chorusGroup.setBounds(bounds.getWidth() - 360, y + 35, 300, height / 2 - 10);
-    m_delayGroup.setBounds(10, y + 35, m_chorusGroup.getX() - 20, height / 2 - 10);
-	m_outputGroup.setBounds(bounds.getWidth() - 360, height / 2 + 45, 300, height / 2 - 30);
-	m_filterGroup.setBounds(10, height / 2 + 45, m_outputGroup.getX() - 20, height / 2 - 30);
-
+#endif
+    m_chorusGroup.setBounds(bounds.getWidth() - 360, y + 35, 300, 165);
+    m_delayGroup.setBounds(10, y + 35, m_chorusGroup.getX() - 20, 165);
+	m_outputGroup.setBounds(bounds.getWidth() - 360, m_delayGroup.getBottom() + 10, 300, 155);
+	m_filterGroup.setBounds(10, m_delayGroup.getBottom() + 10, m_outputGroup.getX() - 20, 265);
+	m_filterModGroup.setBounds(0, m_lowCutFreqKnob.getBottom() + 30, m_filterGroup.getWidth(), 120);
+	
 	m_delayTimeLKnob.setTopLeftPosition(20, 20);
 	m_delayTimeRKnob.setTopLeftPosition(m_delayTimeLKnob.getRight() + 20, 20);
 	m_delayNoteLKnob.setTopLeftPosition(m_delayTimeLKnob.getX(), m_delayTimeLKnob.getY());
@@ -99,14 +130,36 @@ void ViiveAudioProcessorEditor::resized()
 	m_tempoSyncRButton.setCentrePosition(m_delayTimeRKnob.getRight() - 35, m_delayTimeRKnob.getBottom() + 15);
 	m_feedbackKnob.setTopLeftPosition(m_delayTimeRKnob.getRight() + 20, 20);
 	
-	m_delayModeLabel.setTopLeftPosition(10, y);
-	m_modeSelector.setTopLeftPosition(95, y);
+	m_modeSelector.setTopLeftPosition(10, y);
 
 	m_offsetKnob.setTopLeftPosition(m_feedbackKnob.getRight() + 20, 20);
 	m_lowCutFreqKnob.setTopLeftPosition(20, 20);
 	m_lowCutQKnob.setTopLeftPosition(m_lowCutFreqKnob.getRight() + 20, 20);
+	m_lowCutModRateKnob.setTopLeftPosition(20, 20);
+	m_lowCutModNoteKnob.setTopLeftPosition(m_lowCutModRateKnob.getX(), m_lowCutModRateKnob.getY());
+	m_lowCutModDepthKnob.setTopLeftPosition(m_lowCutModRateKnob.getRight() + 10, 20);
+	m_lowCutModPhaseKnob.setTopLeftPosition(m_lowCutModDepthKnob.getRight() + 10, 20);
+	m_lowCutModTempoSyncButton.setTopLeftPosition(20, m_lowCutModRateKnob.getBottom() + 5);
+	m_lowCutModShapeSelector.setTopLeftPosition(m_lowCutModDepthKnob.getX(), m_lowCutModTempoSyncButton.getY());
+
+	constexpr int indicatorSize = 8;
+	m_lowCutLfoIndicator.setSize(indicatorSize, indicatorSize);
+	m_lowCutLfoIndicator.setCentrePosition(m_lowCutModPhaseKnob.getBounds().getCentreX(),
+	                                        m_lowCutModPhaseKnob.getBottom() + 15);
 	m_highCutFreqKnob.setTopLeftPosition(m_lowCutQKnob.getRight() + 20, 20);
 	m_highCutQKnob.setTopLeftPosition(m_highCutFreqKnob.getRight() + 20, 20);
+	m_highCutModPhaseKnob.setTopRightPosition(m_highCutQKnob.getRight(), 20);
+	m_highCutModDepthKnob.setTopRightPosition(m_highCutModPhaseKnob.getX() - 10, 20);
+	m_highCutModRateKnob.setTopRightPosition(m_highCutModDepthKnob.getX() - 10, 20);
+	m_highCutModNoteKnob.setTopLeftPosition(m_highCutModRateKnob.getX(), m_highCutModRateKnob.getY());
+	m_highCutModTempoSyncButton.setTopLeftPosition(m_highCutModRateKnob.getX(), m_highCutModRateKnob.getBottom() + 5);
+	m_highCutModShapeSelector.setTopLeftPosition(m_highCutModDepthKnob.getX(), m_highCutModTempoSyncButton.getY());
+
+	m_highCutLfoIndicator.setSize(indicatorSize, indicatorSize);
+	m_highCutLfoIndicator.setCentrePosition(m_highCutModPhaseKnob.getBounds().getCentreX(),
+	                                         m_highCutModPhaseKnob.getBottom() + 15);
+	m_highCutModInvertButton.setCentrePosition(m_highCutLfoIndicator.getBounds().getCentreX() + 25,
+	                                            m_highCutLfoIndicator.getBounds().getCentreY());
 	m_stereoKnob.setTopLeftPosition(20, 20);
 	m_mixKnob.setTopLeftPosition(m_stereoKnob.getRight() + 20, 20);
 	m_gainKnob.setTopLeftPosition(m_mixKnob.getRight() + 20, 20);
@@ -116,29 +169,34 @@ void ViiveAudioProcessorEditor::resized()
 
 	m_meter.setBounds(m_outputGroup.getRight() + 15, y + 45, 35, height - 45);
 
-
+#if JUCE_DEBUG
+	m_lfoScope->setBounds(10, 490, bounds.getWidth() - 20, 170);
+#endif
 }
 
-void ViiveAudioProcessorEditor::updateDelayKnobs(bool syncLActive, bool syncRActive)
+void ViiveAudioProcessorEditor::updateSyncedKnobs(bool syncLActive, bool syncRActive, bool lowCutSyncActive, bool highCutSyncActive)
 {
 	m_delayTimeLKnob.setVisible(!syncLActive);
 	m_delayNoteLKnob.setVisible(syncLActive);
 	m_delayTimeRKnob.setVisible(!syncRActive);
 	m_delayNoteRKnob.setVisible(syncRActive);
+	m_lowCutModRateKnob.setVisible(!lowCutSyncActive);
+	m_lowCutModNoteKnob.setVisible(lowCutSyncActive);
+	m_highCutModRateKnob.setVisible(!highCutSyncActive);
+	m_highCutModNoteKnob.setVisible(highCutSyncActive);
+	DBG("Updated synced knobs: " << (syncLActive ? "Sync L active, " : "Sync L inactive, ") << (syncRActive ? "Sync R active, " : "Sync R inactive, ") << (lowCutSyncActive ? "Low Cut Sync active, " : "Low Cut Sync inactive, ") << (highCutSyncActive ? "High Cut Sync active" : "High Cut Sync inactive"));
+
 }
 
 void ViiveAudioProcessorEditor::parameterChanged(const juce::String& paramID, float newValue) {
-	if (paramID == tempoSyncLParamID.getParamID() || paramID == tempoSyncRParamID.getParamID()) {
+	if (paramID == tempoSyncLParamID.getParamID() || paramID == tempoSyncRParamID.getParamID() || paramID == lowCutModTempoSyncParamID.getParamID() || paramID == highCutModTempoSyncParamID.getParamID()){
 		juce::MessageManager::callAsync([this] {
 			bool syncL = m_audioProcessor.apvts.getParameter(tempoSyncLParamID.getParamID())->getValue() > 0.5f;
 			bool syncR = m_audioProcessor.apvts.getParameter(tempoSyncRParamID.getParamID())->getValue() > 0.5f;
-			updateDelayKnobs(syncL, syncR);
+			bool lowCutSync = m_audioProcessor.apvts.getParameter(lowCutModTempoSyncParamID.getParamID())->getValue() > 0.5f;
+			bool highCutSync = m_audioProcessor.apvts.getParameter(highCutModTempoSyncParamID.getParamID())->getValue() > 0.5f;
+			updateSyncedKnobs(syncL, syncR, lowCutSync, highCutSync);
 			});
 	}
 }
 
-void ViiveAudioProcessorEditor::modeSelectorChanged() {
-	int selectedID = m_modeSelector.getSelectedId();
-	auto* param = dynamic_cast<juce::AudioParameterChoice*>(m_audioProcessor.apvts.getParameter(delayModeParamID.getParamID()));
-	param->setValueNotifyingHost(param->convertTo0to1(selectedID - 1));
-}