+++ Source/PluginProcessor.cpp
@@ -116,6 +116,8 @@ void ViiveAudioProcessor::processBlock(juce::AudioBuffer<float>& buffer, juce::M
 
     m_tempo.update(getPlayHead());
     m_params.update(m_tempo);
+    if (m_tempo.getIsPlaying())
+        m_delayEngine.updateSyncedModulators(m_tempo.getPpqPosition());
     m_delayEngine.setDelayMode(m_params.delayMode());
 
     //float syncedTimeL = static_cast<float>(m_tempo.noteLengthToMs(m_params.delayNoteL()));